trying to make gps direction correct

+++ mcp/tx.c
@@ -1335,7 +1335,7 @@ static void StoreData(int index)
   WriteData(dgpsLonAddr, (int)(DGPSPos[i_dgps].lon * DEG2I), NIOS_QUEUE);
   WriteData(dgpsAltAddr, DGPSPos[i_dgps].alt, NIOS_QUEUE);
   WriteData(dgpsSpeedAddr, DGPSPos[i_dgps].speed*100, NIOS_QUEUE);
-  WriteData(dgpsDirAddr, DGPSPos[i_dgps].direction*DEG2I, NIOS_QUEUE);
+  WriteData(dgpsDirAddr, (unsigned int)DGPSPos[i_dgps].direction*DEG2I, NIOS_QUEUE);
   WriteData(dgpsClimbAddr, DGPSPos[i_dgps].climb*100, NIOS_QUEUE);
   WriteData(dgpsNSatAddr, DGPSPos[i_dgps].n_sat, NIOS_QUEUE);
